+++ history.md
@@ -1,5 +1,6 @@
 ## 3.2.0 / 2015-11-15
 
+- jss-debug is deprecated, rule name is inside of gemerated class name now
 - no more "type" option for style sheet, its always set to text/css
 - no more "title" option for style sheet, its not a part of the spec http://www.w3.org/TR/html-markup/style.html
 - added meta option, now you can for e.g. set the name of your component and find it on style element.